Integrate OpenAI helpers into adapter
